+++ .github/workflows/generate_artwork.R
@@ -9,7 +9,7 @@ seed <- as.numeric(Sys.Date())
 set.seed(seed)
 
 # Select artwork type
-type <- sample(1:26, size = 1)
+type <- sample(1:27, size = 1)
 
 # Create artwork with random palette, feel free to suggest a new pallette at https://github.com/koenderks/aRtsy/issues
 artwork <- switch(type,
@@ -38,7 +38,8 @@ artwork <- switch(type,
   "23" = canvas_recaman(colors = colorPalette("random-palette"), background = colorPalette("random", 1), iterations = sample(100:1000, size = 1), curvature = sample(1:15, size = 1), start = sample(1:100, size = 1), angle = sample(c(0, 45), size = 1)),
   "24" = canvas_phyllotaxis(colors = colorPalette("random-palette"), iterations = sample(1000:100000, size = 1), p = runif(1, 0.5, 0.9), background = colorPalette("random", 1), angle = runif(1, 0, 1000), size = 0.01, alpha = runif(1, 0.3, 1)),
   "25" = canvas_cobweb(colors = colorPalette("random-palette"), background = colorPalette("random", 1), lines = sample(500:1000, size = 1), iterations = sample(20:100, size = 1)),
-  "26" = canvas_chladni(colors = colorPalette("random-palette"), waves = sample(3:10, size = 1), resolution = sample(c(500, 1000), size = 1), warp = runif(1, 0, 1.5))
+  "26" = canvas_chladni(colors = colorPalette("random-palette"), waves = sample(3:10, size = 1), resolution = sample(c(500, 1000), size = 1), warp = runif(1, 0, 1.5)),
+  "27" = canvas_petri(colors = colorPalette("random-palette"), attractors = 5000, hole = sample(c(0, 0.25, 0.5, 0.8), size = 1))
 )
 
 saveCanvas(artwork, filename, width = ifelse(type == 19, yes = NA, no = 7), height = ifelse(type == 19, yes = NA, no = 7))